Update Rework Inventory
-> continué du rework de l'inventaire
 -> Meilleure gestion des objets dans son ensemble
-> Meilleure séparation des classe et leur responsabilitées
-> ajout de la roadmap de l'inventaire

+++ Assets/Gameplay/Items/Inventory/IInventoryStackable.cs
@@ -6,7 +6,7 @@ namespace LightHouse.Inventory
     public interface IInventoryStackable
     {
         public IInventoryItem Item { get; }
-        public int MaxStack { get; }
+        public ushort MaxStack { get; }
         public int CurrentStack { get; set; }
         public bool CanStackWith(IInventoryItem otherItem);
         public int AddToStack(int amount);
